code/requirements.txt
asn1crypto==1.4.0
attrdict==2.0.1
attrs==20.3.0
blake2b-py==0.1.4
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
coincurve==15.0.0
coloredlogs==15.0
configparser==5.0.2
coverage==5.5
cycler==0.10.0
cytoolz==0.11.0
dictionaries==0.0.2
eth-abi==1.3.0
eth-account==0.3.0
eth-bloom==1.0.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-tester==0.1.0b32
eth-typing==2.2.2
eth-utils==1.9.0
ethereum==2.3.2
ethereum-input-decoder==0.2.2
etherscan-python==2.1.0
future==0.18.2
hexbytes==0.2.1
humanfriendly==9.1
idna==2.10
inflection==0.5.1
Jinja2==2.11.3
jsonschema==3.2.0
kiwisolver==1.3.1
lru-dict==1.1.7
MarkupSafe==1.1.1
matplotlib==3.4.1
mock==4.0.3
mypy-extensions==0.4.3
mythril==0.22.18
mythx-models==2.0.1
numpy==1.19.0
parsimonious==0.8.1
pbkdf2==1.3
persistent==4.6.4
Pillow==8.2.0
plyvel==1.3.0
psycopg2==2.8.6
py-ecc==1.6.0
py-evm==0.3.0a13
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.0.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
pyethash==0.1.27
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pysha3==1.0.2
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.7.1
PyYAML==5.4.1
repoze.lru==0.7
requests==2.25.0
rlp==1.2.0
scrypt==0.8.17
semantic-version==2.8.5
six==1.15.0
toolz==0.11.1
transaction==3.0.1
trie==1.4.0
urllib3==1.26.4
z3-solver==4.8.10.0
zope.interface==5.3.0
